html/testdata/go/fostered_foreign_content.dat

#data
<math><mtext><table><mglyph><style><img>
#errors
#document
| <html>
|   <head>
|   <body>
|     <math math>
|       <math mtext>
|         <mglyph>
|           <style>
|             "<img>"
|         <table>
